(2) v0.3.0 Level 3 Start
A new Level 3 was added, featuring a stage that requires backtracking. This isn't finished yet.

A script to control tutorial controls opacity was added.

Parallax background shaking was fixed.

A new layer for platforms was added, and the player was given a new collider. The old collider can't interact with platforms, but the new one can.