(2) v0.3.1a Level 3 Advances
Level 3 was changed, but not a big enough change to get its own numbered version.

The boss ghost now checks if its health bar is null before trying to update it.

The player now uses extrapolation on the rigidbody, since some devides had stuttering. This needs to be investigated further.