(2) v0.6.0 Instant Color Change
The Player can now select their color instantly rather than scrolling one by one. This is done with the numbers 1-4 on keyboard, or the DPAD with a controller. This means that the dpad is no longer used for moving.

The player color switching is now instant, regardless if they have their lantern active or not.