(2) v0.7.0 Level 4 Start
Level 4 was reset, and severely modified, although only the starting zone is implemented.

A Multi Icon Manager was implemented, which does the same as the normal Icon Manager but for multiple of the same references.

Tutorial Controls Enabler now also has an additional check that allows us to prevent the tutorial icons from appearing until the player has unlocked a certain number of colors.